Reached parsability of the standard corpus file Implemented interpreter for point class declarations

+++ src/plang/lang/logic/manager.py
@@ -3,7 +3,8 @@ from typing import List, Optional, Dict, Union
 from sqlalchemy import inspect
 from sqlalchemy.orm import Session
 
-from plang.error import LangExceptions
+from plang.db.plot.object.models import ObjectClass, ObjectClassHintLit, ObjectClassHintSym, ObjectClassHintPnt
+from plang.db.plot.point.models import PointClass, PointClassHint
 from plang.cli.scope import Scope
 from plang.db.base import Base, FormBase
 from plang.db.models import Path
@@ -27,7 +28,8 @@ class Report:
         sa = 'Added entries:\n\t' + '\n\t'.join([a.str(rich) for a in self.added]) if len(self.added) > 0 else ''
         sm = 'Modified entries:\n\t' + '\n\t'.join([a.str(rich) for a in self.modified]) if len(
             self.modified) > 0 else ''
-        sd = 'Deleted entries:\n\t' + '\n\t'.join([a.str(rich) for a in self.deleted]) if len(self.deleted) > 0 else ''
+        # sd = 'Deleted entries:\n\t' + '\n\t'.join([a.str(rich) for a in self.deleted]) if len(self.deleted) > 0 else ''
+        sd = []
         sf = 'Failed entries:\n\t' + '\n\t'.join([str(a) for a in self.failed]) if len(self.failed) > 0 else ''
         return '\n'.join(s for s in [sa, sm, sd, sf] if len(s) > 0)
 
@@ -54,7 +56,6 @@ class Manager:
         else:
             node = self.scope.getPath()
 
-        leaf = len(form.nodes) - 1
         for i in range(0, len(form.nodes)):
             nodes = [n for n in node.children if n.name == form.nodes[i]]
             if len(nodes) != 1 and insert:
@@ -114,7 +115,8 @@ class Manager:
         return result if result.id is not None else None
 
     def removePath(self, path: Path, recursive: bool = False):
-        self.session.delete(path)
+        if recursive or len(path.children) == 0:
+            self.session.delete(path)
 
     def selectOrInsertSymbolClass(self, form: SymbolClass.Form, insert: bool = True) -> Optional[SymbolClass]:
         path = self.selectPath(form.path) or self.selectOrInsertPath(form.path, insert)
@@ -169,15 +171,15 @@ class Manager:
         self.session.delete(symbol_class)
 
     def selectOrInsertSymbol(self, form: Symbol.Form, insert: bool = True) -> Optional[Symbol]:
-        symbolClass = self.selectOrInsertSymbolClass(form.clazz, insert)
-        if symbolClass is None:
+        symbol_class = self.selectOrInsertSymbolClass(form.clazz, insert)
+        if symbol_class is None:
             return None
         else:
-            symbol = next(filter(lambda x: x.name == form.name, symbolClass.instances), None)
+            symbol = next(filter(lambda x: x.name == form.name, symbol_class.instances), None)
             if symbol is None:
                 symbol = Symbol()
                 symbol.name = form.name
-                symbol.clazz = symbolClass
+                symbol.clazz = symbol_class
                 self.session.add(symbol)
 
             if form.decoration is not None:
@@ -204,4 +206,103 @@ class Manager:
         pass
 
     def removeSymbol(self, symbol: Symbol):
+        self.session.delete(symbol)
+
+    def selectObjectClass(self, form: ObjectClass.Form) -> Optional[ObjectClass]:
         pass
+
+    def insertObjectClass(self, form: ObjectClass.Form) -> Optional[ObjectClass]:
+        path = self.selectPath(form.path_form) or self.selectOrInsertPath(form.path_form, True)
+        point_class = path.point_class
+        if point_class is None:
+            return None
+        object_class = next(filter(lambda x: x.name == form.name, point_class.objects), None)
+        if object_class is None:
+            object_class = ObjectClass()
+            object_class.name = form.name
+            object_class.point_class = point_class
+            self.session.add(object_class)
+            self.__refresh(point_class)
+        object_class.default = form.default
+        object_class.singleton = form.singleton
+        for h in object_class.hints_lit:
+            self.session.delete(h)
+        for h in object_class.hints_sym:
+            self.session.delete(h)
+        for h in object_class.hints_pnt:
+            self.session.delete(h)
+        for h in form.hint_list:
+            pass
+            if isinstance(h, ObjectClassHintLit.Form):
+                lit_hint = ObjectClassHintLit()
+                lit_hint.type = h.type
+                lit_hint.hint = str(h.hint)
+                lit_hint.clazz = object_class
+                self.session.add(lit_hint)
+            elif isinstance(h, SymbolClass.Form):
+                sym_hint = ObjectClassHintSym()
+                sym_hint.hint = self.selectOrInsertSymbolClass(h)
+                sym_hint.recursive = h.recursive
+                sym_hint.clazz = object_class
+                self.session.add(sym_hint)
+            elif isinstance(h, PointClass.Form):
+                pnt_hint = ObjectClassHintPnt()
+                pnt_hint.hint = self.selectOrInsertPointClass(h)
+                pnt_hint.recursive = h.recursive
+                pnt_hint.clazz = object_class
+                self.session.add(pnt_hint)
+            self.__refresh(object_class)
+
+        return object_class
+
+    def removeObjectClass(self, object_class: ObjectClass):
+        self.session.delete(object_class)
+
+    def selectOrInsertPointClass(self, form: PointClass.Form, insert: bool = True) -> PointClass:
+        return self.selectPointClass(form) or self.insertPointClass(form)
+
+    def selectPointClass(self, form: PointClass.Form) -> Optional[PointClass]:
+        path = self.selectPath(form.path_form)
+        if path is None:
+            return None
+        return path.point_class
+
+    def insertPointClass(self, form: PointClass.Form) -> Optional[PointClass]:
+        path = self.selectPath(form.path_form) or self.selectOrInsertPath(form.path_form, True)
+        if form.path_form.decoration is not None:
+            path.ordinal = form.path_form.decoration.ordinal
+            path.description = form.path_form.decoration.description
+        if path.point_class is None:
+            point_class = PointClass()
+            point_class.path = path
+            self.session.add(point_class)
+            self.__refresh(path)
+        else:
+            point_class = path.point_class  # TODO: rest
+
+        point_class.singleton = form.singleton
+        if len(point_class.hints) > 0:
+            for h in point_class.hints:
+                self.session.delete(h)
+            self.__refresh(point_class)
+        for h in form.hint_list:
+            point_class_hint = PointClassHint()
+            point_class_hint.hint = self.selectOrInsertSymbolClass(h)
+            point_class_hint.clazz = point_class
+            point_class_hint.recursive = h.recursive
+            self.session.add(point_class_hint)
+
+        if len(point_class.objects) > 0:
+            for o in point_class.objects:
+                self.session.delete(o)
+            self.__refresh(point_class)
+        if len(form.object_class_forms) > 0:
+            point_class_scope = self.scope.getScope(path)
+            object_class_manager = Manager(self.session, point_class_scope)
+            for o in form.object_class_forms:
+                object_class_manager.insertObjectClass(o)
+
+        return point_class
+
+    def removePointClass(self, point_class: PointClass):
+        self.session.delete(point_class)
